refactor: add response/request related definitions in respective modules
