+++ decaf-client.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 41924fa1a6d538fc3cad9c387b6199502e3a66a6e301eddd36e5c458886520b2
+-- hash: 90f836bda379684c199708d787f74e3d2e29bf5ada683fe6f9fdd0e73ae4a518
 
 name:           decaf-client
 version:        0.0.0.3
@@ -33,11 +33,11 @@ library
       Decaf.Client.Internal.Apis.Barista
       Decaf.Client.Internal.Apis.Microlot
       Decaf.Client.Internal.Apis.Pdms
-      Decaf.Client.Internal.Combinators
       Decaf.Client.Internal.Error
       Decaf.Client.Internal.Http
       Decaf.Client.Internal.Remote
       Decaf.Client.Internal.Request
+      Decaf.Client.Internal.Response
       Decaf.Client.Internal.Types
       Decaf.Client.Internal.Utils
       Decaf.Client.Version
