+++ src/Decaf/Client/Internal/Types.hs
@@ -5,110 +5,9 @@
 
 module Decaf.Client.Internal.Types where
 
-import qualified Data.ByteString.Lazy         as BL
-import qualified Data.Text                    as T
-import           Decaf.Client.Internal.Remote (Remote)
-import           Decaf.Client.Internal.Utils  (dropTrailing)
-import qualified Deriving.Aeson               as DA
-import qualified Deriving.Aeson.Stock         as DAS
-
-
--- | Type definition for high-level encoding of DECAF client requests.
-data Request = Request
-  { requestRemote            :: !Remote
-  , requestNamespace         :: !Path
-  , requestCredentials       :: !Credentials
-  , requestUserAgent         :: !T.Text
-  , requestHttpHeaders       :: !Headers
-  , requestHttpMethod        :: !Method
-  , requestHttpPath          :: !Path
-  , requestHttpTrailingSlash :: !Bool
-  , requestHttpParams        :: ![Param]
-  , requestHttpPayload       :: !(Maybe Payload)
-  }
-
-
-instance Show Request where
-  show x = dropTrailing '\n' $ unlines
-    [ "Request {"
-    , "  requestRemote            = " <> show (requestRemote x)
-    , "  requestNamespace         = " <> show (requestNamespace x)
-    , "  requestCredentials       = " <> show (requestCredentials x)
-    , "  requestUserAgent         = " <> show (requestUserAgent x)
-    , "  requestHttpHeaders       = " <> show (requestHttpHeaders x)
-    , "  requestHttpMethod        = " <> show (requestHttpMethod x)
-    , "  requestHttpPath          = " <> show (requestHttpPath x)
-    , "  requestHttpTrailingSlash = " <> show (requestHttpTrailingSlash x)
-    , "  requestHttpParams        = " <> show (requestHttpParams x)
-    , "  requestHttpPayload       = " <> show (requestHttpPayload x)
-    , "}"
-    ]
-
-
--- | Data definition for DECAF API response values.
-data Response a = Response
-  { responseStatus  :: !Int
-  , responseHeaders :: Headers
-  , responseValue   :: !a
-  } deriving Show
-
-
--- | Type definition for DECAF client request header key/value pair.
-type Header = (T.Text, T.Text)
-
-
--- | Type definition for a list of DECAF client request 'Header's.
-type Headers = [Header]
-
-
--- | Type definition for DECAF client request querystring parameter.
-type Param = (T.Text, T.Text)
-
-
--- | Type definition for a list of DECAF client request 'Param's.
-type Params = [Param]
-
-
--- | Type definition for a list of DECAF client request HTTP path segments.
-newtype Path = MkPath { unPath :: [T.Text] } deriving Show
-
-
--- >>> mkPath "/a/b" <> mkPath "/c/d"
--- MkPath {unPath = ["a","b","c","d"]}
-instance Semigroup Path where
-  (<>) (MkPath p1) (MkPath p2) = MkPath (p1 <> p2)
-
-
--- >>> mconcat [mkPath "/a/b", mkPath "/c/d"]
--- MkPath {unPath = ["a","b","c","d"]}
-instance Monoid Path where
-  mempty = MkPath []
-  mappend = (<>)
-  mconcat ps = MkPath $ concatMap unPath ps
-
-
--- | Sanitizes the given 'T.Text' into path segments and builds a 'Path' value.
---
--- >>> mkPath ""
--- MkPath {unPath = []}
--- >>> mkPath "/"
--- MkPath {unPath = []}
--- >>> mkPath "//"
--- MkPath {unPath = []}
--- >>> mkPath "/a"
--- MkPath {unPath = ["a"]}
--- >>> mkPath "a/"
--- MkPath {unPath = ["a"]}
--- >>> mkPath "/a/"
--- MkPath {unPath = ["a"]}
--- >>> mkPath "/a/b"
--- MkPath {unPath = ["a","b"]}
--- >>> mkPath "a/b/"
--- MkPath {unPath = ["a","b"]}
--- >>> mkPath "/a/b/"
--- MkPath {unPath = ["a","b"]}
-mkPath :: T.Text -> Path
-mkPath = MkPath . filter ("" /=) . T.split ('/' ==)
+import qualified Data.Text            as T
+import qualified Deriving.Aeson       as DA
+import qualified Deriving.Aeson.Stock as DAS
 
 
 -- | Data definition for available DECAF credentials types.
@@ -162,18 +61,3 @@ data KeyCredentials = KeyCredentials
   }
   deriving (Eq, DA.Generic)
   deriving (DA.FromJSON, DA.ToJSON) via DAS.PrefixedSnake "keyCredentials" KeyCredentials
-
-
--- | Available DECAF endpoint methods.
-data Method = GET | POST | PUT | DELETE deriving (Show)
-
-
--- | Data definition for DECAF client request payloads.
-data Payload = Payload
-  { payloadType    :: !T.Text         -- ^ HTTP content type.
-  , payloadContent :: !BL.ByteString  -- ^ HTTP payload body.
-  }
-
-
-instance Show Payload where
-  show _ = "<TRUNCATED>"
